+++ nogotofail/mitm/connection/handlers/data/ssl.py
@@ -19,7 +19,10 @@ from nogotofail.mitm.connection.handlers.data import handlers
 from nogotofail.mitm.connection.handlers.data import DataHandler
 from nogotofail.mitm.connection.handlers.store import handler
 from nogotofail.mitm.event import connection
+from nogotofail.mitm import util
 from nogotofail.mitm.util import ssl2, tls, vuln
+from nogotofail.mitm.util.tls.types import HandshakeMessage, TlsRecord
+
 
 @handler.passive(handlers)
 class InsecureCipherDetectionHandler(DataHandler):
@@ -49,7 +52,7 @@ class InsecureCipherDetectionHandler(DataHandler):
                 (", ".join(null_ciphers)))
 
         # Check for NULL integrity ciphers
-        integ_ciphers =  [str(c) for c in client_hello.ciphers if str(c).endswith("_NULL")]
+        integ_ciphers = [str(c) for c in client_hello.ciphers if str(c).endswith("_NULL")]
         if integ_ciphers:
             self._handle_bad_ciphers(integ_ciphers,
                 "Client enabled NULL integrity TLS/SSL cipher suites %s" %
@@ -90,3 +93,71 @@ class WeakTLSVersionDetectionHandler(DataHandler):
                 self.log(logging.ERROR,
                         "Client enabled SSLv3 protocol without TLS_FALLBACK_SCSV")
             self.log_attack_event(data="SSLv3")
+
+
+@handler.passive(handlers)
+class NoForwardSecrecy(DataHandler):
+    name = "noforwardsecrecy"
+    description = (
+        "Detects selected server cipher suites which don't support "
+        "Diffie-Hellman key exchange (DHE or ECDHE) i.e. in "
+        "SERVER_HELLO response")
+
+    buffer = ""
+    ssl = False
+
+    def on_ssl(self, client_hello):
+        self.ssl = True
+        self.client_session_id = client_hello.session_id
+        return False
+
+    def on_response(self, response):
+        if not self.ssl:
+            return response
+        response = self.buffer + response
+        self.buffer = ""
+        try:
+            index = 0
+            while index < len(response):
+                record, size = TlsRecord.from_stream(response[index:])
+                for i, message in enumerate(record.messages):
+                    # Check for Server Hello message
+                    if (isinstance(message, tls.types.HandshakeMessage) and
+                            message.type == HandshakeMessage.TYPE.SERVER_HELLO):
+                        server_hello = message.obj
+                        selected_cipher = str(server_hello.cipher)
+
+                        _connection = self.connection
+                        destination = _connection.hostname if \
+                            _connection.hostname else _connection.server_addr
+                        debug_message = ["Selected cipher \"", selected_cipher,
+                                "\" for connection to \"", destination, "\""]
+                        self.log(logging.DEBUG, "".join(debug_message))
+                        """ Check if Ephemeral Diffie-Hellman key exchange is
+                            used in selected cipher """
+                        fs_key_strings = ["DHE", "ECDHE"]
+                        if not [fs_string for fs_string in fs_key_strings
+                                if fs_string in selected_cipher]:
+                            error_message = \
+                                ["Cipher suite key exhange technqiue doesn't ",
+                                 "support forward secrecy. ",
+                                 "Cipher suite - [", selected_cipher, "]"]
+                            self.log(logging.WARNING, "".join(error_message))
+                            self.log_event(logging.WARNING,
+                                connection.AttackEvent(
+                                    self.connection, self.name, True, ""))
+                            self.connection.vuln_notify(
+                                util.vuln.VULN_NO_FORWARD_SECRECY)
+                        return response
+                index += size
+        except ValueError:
+            # Failed to parse TLS, this is probably due to a short read of a TLS
+            # record. Buffer the response to try and get more data.
+            self.buffer = response
+            # But don't buffer too much, give up after 16k.
+            if len(self.buffer) > 2**14:
+                response = self.buffer
+                self.buffer = ""
+                return self.buffer
+            return ""
+        return response
